Use bigint

+++ src/main/kotlin/no/nav/hjelpemidler/soknad/db/domain/SoknadMedStatus.kt
@@ -1,12 +1,13 @@
 package no.nav.hjelpemidler.soknad.db.domain
 
 import com.fasterxml.jackson.databind.JsonNode
+import java.math.BigInteger
 import java.util.Date
 import java.util.UUID
 
 class SoknadMedStatus private constructor(
     val soknadId: UUID,
-    val journalpostId: ULong?,
+    val journalpostId: BigInteger?,
     val datoOpprettet: Date,
     val datoOppdatert: Date,
     val status: Status,
@@ -15,10 +16,10 @@ class SoknadMedStatus private constructor(
     val er_digital: Boolean
 ) {
     companion object {
-        fun newSøknadUtenFormidlernavn(soknadId: UUID, journalpostId: ULong?, datoOpprettet: Date, datoOppdatert: Date, status: Status, fullmakt: Boolean, er_digital: Boolean) =
+        fun newSøknadUtenFormidlernavn(soknadId: UUID, journalpostId: BigInteger?, datoOpprettet: Date, datoOppdatert: Date, status: Status, fullmakt: Boolean, er_digital: Boolean) =
             SoknadMedStatus(soknadId, journalpostId, datoOpprettet, datoOppdatert, status, fullmakt, null, er_digital)
 
-        fun newSøknadMedFormidlernavn(soknadId: UUID, journalpostId: ULong?, datoOpprettet: Date, datoOppdatert: Date, status: Status, fullmakt: Boolean, søknad: JsonNode, er_digital: Boolean) =
+        fun newSøknadMedFormidlernavn(soknadId: UUID, journalpostId: BigInteger?, datoOpprettet: Date, datoOppdatert: Date, status: Status, fullmakt: Boolean, søknad: JsonNode, er_digital: Boolean) =
             SoknadMedStatus(soknadId, journalpostId, datoOpprettet, datoOppdatert, status, fullmakt, formidlerNavn(søknad), er_digital)
     }
 }
